Fixed sorting function
Sorting function originally sorted column values as strings. For example, "900" is bigger than "1000", because in the first character, 9>1.
This is obviously undesirable behavior if we want to sort them as numbers, where 1000>900.
I added a condition to the sorting function where if all entries in a column are detected to have some number in them, then we convert the strings to integers after stripping out unnecessary characters, then sort the list as integers.
Otherwise, if even a single entry doesn't contain numbers, then the column cannot be sorted by numbers, and we sort as usual as strings.